requirements.txt
streamlit
ultralytics
pandas
numpy
openpyxl
Pillow

